+++ nengo/tests/conftest.py
@@ -190,26 +190,6 @@ def pytest_generate_tests(metafunc):
             "nl_nodirect", [LIF, LIFRate, RectifiedLinear, Sigmoid])
 
 
-def pytest_addoption(parser):
-    parser.addoption(
-        '--plots', nargs='?', default=False, const=True,
-        help='Save plots (can optionally specify a directory for plots).')
-    parser.addoption(
-        '--analytics', nargs='?', default=False, const=True,
-        help='Save analytics (can optionally specify a directory for data).')
-    parser.addoption(
-        '--compare', nargs=2,
-        help='Compare analytics results (specify directories to compare).')
-    parser.addoption(
-        '--logs', nargs='?', default=False, const=True,
-        help='Save logs (can optionally specify a directory for logs).')
-    parser.addoption('--noexamples', action='store_false', default=True,
-                     help='Do not run examples')
-    parser.addoption(
-        '--slow', action='store_true', default=False,
-        help='Also run slow tests.')
-
-
 def pytest_runtest_setup(item):
     for mark, option, message in [
             ('example', 'noexamples', "examples not requested"),